implement !include_dir_merge_recursive
The will merge dicts recursively (if two or more files contain a dict with a mapping that has an identical key, both values will be merged). dicts will be merged by applying the merge algorithm to the values of identical keys. lists will be merged by extending the existing lists. All other types lead to an exception. Recursion is stopped when hitting a value that is not a dict.